+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Added *Change GP Stroke Direction* node.
 - Added create list option to the *Random Quaternion* node.
 - Added *Random Boolean* node.
+- Added *Random Color* node.
+- Added *Find Shortest Path* node.
 
 ### Fixed
 
@@ -85,6 +87,7 @@
 - Added mesh *Triangulation* methods and *Triangulate Mesh* node.
 - Added Mesh Points Scatter node.
 - Added support for 2D vectors in 3D viewer node.
+- Added *Vertex Color Fill* in *GP Stroke Info*, *GP Stroke From Points*, and *Set GP Stroke Attributes* nodes.
 
 ### Fixed